requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
Faker==13.12.0
gunicorn==20.1.0
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
